fix: resolve num-to-double/int clamp type errors across flutter app

+++ ungraey_flutter/lib/features/bounties/ui/screens/create_bounty_view.dart
@@ -35,7 +35,7 @@ class _CreateBountyModalState extends ConsumerState<CreateBountyModal> {
     final draft = widget.draft;
     if (draft != null) {
       _category = draft.category;
-      _rewardDollars = draft.rewardDollars.clamp(1, 100);
+      _rewardDollars = draft.rewardDollars.clamp(1, 100).toInt();
       _titleController.text = draft.titleHint;
     }
   }
